fix redundant datastores issue
