--- app/src/main/java/com/giraffe/tudeeapp/domain/service/MainService.kt
@@ -1,6 +1,0 @@
-package com.giraffe.tudeeapp.domain.service
-
-interface MainService {
-    suspend fun setCurrentTheme(isDark: Boolean)
-    suspend fun getCurrentTheme(): Boolean
-}
