+++ app/src/main/java/com/giraffe/tudeeapp/domain/service/TudeeAppService.kt
@@ -1,6 +1,9 @@
 package com.giraffe.tudeeapp.domain.service
 
-interface SplashService {
+interface TudeeAppService {
+    suspend fun setCurrentTheme(isDark: Boolean)
+    suspend fun getCurrentTheme(): Boolean
+
     suspend fun setOnboardingShown(shown: Boolean)
     suspend fun isOnboardingShown(): Boolean
 }
