+++ app/src/main/java/com/giraffe/tudeeapp/presentation/splash/viewmodel/SplashViewModel.kt
@@ -2,13 +2,13 @@ package com.giraffe.tudeeapp.presentation.splash.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.giraffe.tudeeapp.domain.service.SplashService
+import com.giraffe.tudeeapp.domain.service.TudeeAppService
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 
 class SplashViewModel(
-    private val splashService: SplashService
+    private val appService: TudeeAppService
 ) : ViewModel() {
 
     private val _isOnboardingShown = MutableStateFlow<Boolean?>(null)
@@ -16,13 +16,13 @@ class SplashViewModel(
 
     fun checkOnboardingStatus() {
         viewModelScope.launch {
-            _isOnboardingShown.value = splashService.isOnboardingShown()
+            _isOnboardingShown.value = appService.isOnboardingShown()
         }
     }
 
     fun setOnboardingShown(shown: Boolean) {
         viewModelScope.launch {
-            splashService.setOnboardingShown(shown)
+            appService.setOnboardingShown(shown)
         }
     }
 }
